TASK: Use http host instead of get parameter for docs rendering
As we have a custom domain just for the documentation rendering hook, we
detect requests to render documentation by checking for the domain.

This removes the need for a special get argument. Also we are way more
flexible as we could replace intercept by something else on the domain.

Right now the domain is an cname to intercept.

Relates: DOCS-1