[TASK] Add a testing related double bundle and use in tests
Symfony does not easily allow substituting single (third-party)
dependencies in functional tests. One reason is that each request
creates a new kernel and container, which is obvious for
$client->request(), but also happens implicitly if a test does
something like $form->submit(). Tests like these do multiple
requests: First one to render a form, second to submit form data.

If such a form submit triggers a third party call - in our case
to bamboo for example, a test should be able to mock that call.
But that's not easily possible. Searching for this a casual answer
is "you're doing it wrong" and those who ask for the valid case
that for instance a guzzle call to a third party should be mockable
are not answered.

This solution is inspired by
https://github.com/docteurklein/TestDoubleBundle
but allows injecting object prophecies prepared by the test itself
for as many calls as needed AND having the prophecy available in
the test directly.