[TASK] Add status to DocumentationJar so it can be locked for actions
This will add a isActionable method to the Jar. You can call this to see
if this Jar can be delete or re-rendered etc. This is so that we do not
re-render a Jar when it has been deleted a few seconds ago. This also
prevents double rendering (needs to be implemented separately)