[BUGFIX] Drop support for "release" events from Github webhook
When a release is created on Github, multiple requests are sent to
webhooks if configured as "Sends me everything". All requests contain
the header X-GitHub-Event, but only one of these requests has its
value set to push, all other requests are headed with release.

Since the documentation says that a webhook should be configured for
push events only, it's considered safe to remove the release
support.

Resolves: DEV-394

Squashed commit of the following:

commit 9da829f15cf853597b3d6712c87cc537655681d6
Author: Andreas Fernandez <andreas.fernandez@typo3.com>
Date:   Wed Nov 6 14:16:58 2019 +0100

    [BUGFIX] Drop support for "release" events from Github webhook

    When a release is created on Github, multiple requests are sent to
    webhooks if configured as "Sends me everything". All requests contain
    the header `X-GitHub-Event`, but only one of these requests has its
    value set to `push`, all other requests are headed with `release`.

    Since the documentation says that a webhook should be configured for
    `push` events only, it's considered safe to remove the `release`
    support.

    Resolves: DEV-394