Add redirects for "TypoScript Syntax"
This manual was merged into TYPO3 Explained.

Because we cannot guarantee that all file names stayed the same
and some of it was moved to TSref, we propose a simplified solution:

- redirect everything to the start page in TYPO3 Explained

(The same procedure must later be done for TYPO3 FAL, TYPO3 inside, CGL,
Services etc.). This cannot be done yet, as the merging has not beem
performed yet for versions < 9.5.