[TASK] bamboo-specs: inital YAML specs
The goal is to simplify the bamboo build setup and reduce the overhead
that is currently required to build and deploy the build config(s).

This commit introduces YAML config and external shell scripts.

By using YAML we improve readability and make the configs easier to
maintain. For example there is no need for Java & Maven anymore.

By using external scripts we decouple the build logic from Bamboo and
reduce the amount of logic inside each build config.

The scripts are now easier to test and don't clog the build config.

Ticket(s): INFRA-62