[BUGFIX] Trigger docs rendering for standalone files
Make sure that rendering documentation is triggered for standalone
README files as well.

Currently, supplying documentation in a dedicated Documentation
directory is supported, but also using standalone README.md or
README.rst files without the Documentation directory.

For extensions, this has the advantage that the documentation
is visible immediatedly when visiting the project on the
git hoster (e.g. GitHub) and may be a better fit for extensions
with only minimal documentation.

In any case, it is recommended to add Documentation/Settings.cfg,
so that the title is generated correctly.

Related: TYPO3-Documentation/T3DocTeam#152