[BUGFIX] Symlink core checkouts for performance reasons (#55)
The core release process works in two different ways:

* Splitter
  The core splitter clones the mono repository and adds each extension
  repository as a new git remote to push to.
* Tagger
  The tagger clones each extension repository into a distinguishable
  location, creates the tags and pushes them.

All of these checkouts are now configured to be in shared/var instead
of var/ of a release and get symlinked instead after deploying a new
release as core checkouts are not related to a specific intercept
release. This also improves the performance as the checkouts don't
need to be done again.