[FEATURE] Add documentation validator
The documentation validator checks the documentation's reST files and
Settings.cfg for compliance with TYPO3 documentation standards and
lists selected errors. The authors identified by composer.json are
notified and the Intercept history log provides the validation errors
for the remaining contributors.

For now, the validation errors are handled gently and documentation
rendering is performed anyway. After a sufficient grace period, strict
validation will be introduced.

- Support resolution of any public repository file path
- Adapt email template to GitHub style
- Support parsing of Settings.cfg