[BUGFIX][DEV-1630] Convert tags to branch names for specific repositories
When registering documentation renderings, extension manuals must follow
specific rules, means that only tags are accepted.
However, there's an exception for some repositories where only branches
are allowed. This collides with the "tag only" rule and leads to an
unresolvable situation.

The validation is now adapted for special repositories that tries to
derive the branch name from the tag name and uses this result instead.