[BUGFIX] Fix potential build key collision
When documentation gets rendered, a key `build_key` meant to be unique
is generated that is used to identify a documentation registration
through the whole build process. I has been discovered that the
`build_key` has a low entropy, consisting of the current time and the
package name whose documentation is rendered only, which is prone to
collisions when multiple commits are pushed in different branches
within the same repository and therefore the according webhook to
Intercept is called multiple times in the same second.

When the documentation rendering succeeds, only the last documentation
of said colliding `build_key` will be picked and marked as "rendered",
leaving any other documentation with the same `build_key` marked as
"rendering".

This issue was introduced on April 9, 2021 with commit
2bb608a7c5fcca222048eb596d0f2667ef29c03c.

The issue has been solved by:
* adding more entropy
  * package name
  * source branch
* switching from `sha1` to `xxh128` hashing algorithm
  * using cryptographically random string of 256 bytes as hash secret