[TASK] Redirect non-handled extensions to TER and remove `typo3cms/exceptions`
This commit updates the nginx redirects configuration to redirect any
extension, that has not been redirected before by the dynamic
configuration to extensions.typo3.org in order to get rid of
`typo3cms/extensions`.

Also, the handling for `typo3cms/exceptions` has been removed, the
directory is archived on the server.

Resolves: https://app.asana.com/0/1206620330068017/1207393774643690